requirements.txt
azure-functions
fastapi
python-dotenv
sqlalchemy
psycopg2-binary